Fix goroutine deadlock in stdin
When I would do like `cat /dev/random | cryptocli dd -to-byte-in 32`
it would panic because of deadlock. This was caused when the byteCounter
was actually closing the writer before stdin would close it.
Moving stdin reader to its own goroutine and keeping the channel logic
from when stdout is done writing fixed the problem.