Add inout Pipe module
In addition to using the std or the file module you can now start a command in
a sub shell. Either stdin or stdout is attached to the shell and the command is
passed to it.
It doesn't work on Windows -- not tested but I don't think so.

You can use it like this: `pipe:"curl http://google.com | cat"`

Everything you have in env will be passed to the sub shell which by default is
what's in your `${SHELL}`. If the environment variable is not found, it uses
`/bin/sh`.