Refact Chomp method
I was never satisfy the way the chomp option was done just to handle one case:
`cryptocli dd` where `std` is selected for out but `stdin` is closed.
So I removed the `bool` parameter from `Init` and created `Chomp` instead where
the flag will be passed to the function when the last output codec is done reading.