Add get-certs command
The get-certs command will establish a TLS handshake with a server without sending
any data.
Then it will PEM encode every certificate it finds and render them available to read.

It's good to remember that the handshake doesn't actually verify the validity of the
certificate in this case, therefor you can get-certs self-signed certificates.