Add pipeline package
The pipeline package creates a pipeline of elements that are of type
io.ReadWriteClose and Init(). The purpose is to connect the input to the output
of each element and thus forming a chain.
It then expose a io.ReadWriteCloser and Init() interface that will pass the data
along the pipeline.