Verify checksum in dgst command.
Added two flags:

  - checksum <string>
  - checksum-decoders ...<string>

When -checksum is specified, it will read the whole string and decode it
with the -checksum-decoders codecs. Then at the end before closing, it will
compare the checksum from the cmd-in to the one specified from the command line.
If the two match, nothing is output and exit(0) is expected.
If not, an error is thrown and exit(1) is expected.

Not specifying -checksum won't change the behavior of dgst.