Refact some of the PipeInput internals
We wraps errors from the goroutine and send that to in.sync to be caught
in Close(). Also we makes sure that the last Read() has been called to
unlock in.sync. Otherwise Close() is called before last Read() and since
we os.Exit(1) when calling Close() if err is thrown, the output is
truncated.