Add support for AES GCM algorithm
2 new commands have been added:

  - aes-gcm-encrypt
  - aes-gcm-decrypt

They both support 128 and 256 bits keys.
Since AES GCM doesn't support streaming, we actually construct the cipher text stream
from multiple blocks.

First we derive the key from a source. If the user wants some salt, it is appended to the output. Once the key is derived it creates an 8 bytes nonce then append it to output.
Once 16KiB or less is read from input, we add a 4 bytes number starting from 0 to the nonce and use the 12 bytes nonce to construct the ciphertext. Until we reach EOF we read repeat reading 16KiB or less, incrementing the last for 4 of the nonce.

Note that the nonce only needs to be written at the beginning because we increment the 4 last byte to make sure the nonce never gets reused.