Add nonce rollover
Since we're using unsigned int, after 0xffffffff it will be 0x00000000
meaning that we're going to have the same nonce as the first ones. This is bad.

It's not a huge concern because you will reach this value in 2**32 * the block size.
By default you'll reuse the nonce after 2**32 * 2**14 so after having transfered
64PiB which is beyond the scope of this project.

But since it is a flaw it needs to be fixed even for theoratical issues.