Refactor the reader.go file
When I designed the reader file, I used it with bufio/readfull. It was not possible
to use it in a stream fashion. Meaning that everything was going to be blocking
until readfull finishes. I then added the line flag but it was foolish.

Simply use the Read() providing a buffer. It will return as soon as the kernel
passes the bytes. Not everything might be to send but we know how much it has
written.

It's way more elegant like this.