Refactor stdin reading
I did not have great success making os.Stdin.Close() as it wouldn't stop reading
from stdin.

So I came up with a plan of letting it read but not sending the message if
the in channel would be closed for some reason. Then not waiting on finishing
reading because it would be closed as soon as the program ended.

In short: let it fail but don't send any message after the pipeline has been closed.