Remove the --data flag in http module
Initial work for the json flag which incidentaly made me realize I could
remove the `--data` flag. It is replaced by adding the `--data-timeout`
flag that will be triggered if no data is received from the pipeline.

Usually, unless it is to mirror data, the pipeline will already be
filled with data when reaching the module.
But HTTP does not work with mirroring data, because the server has to
fully read the request in order to send the response. This why we
have websockets.