New pwn feature
This feature makes it possible to read/write from the pipeline.
You just have to specify the javascript file (can be a pipe), then create
your `start` function.
The pipeline will be blocked until it finish running the `start` function.
It means that you will have to create some kind of a loop, otherwise the
channel is recycled.

There are builtins for you:
  - read(int) string
  - readline() string
  - write(string)
  - toJSON(object) string
  - fromJSON(string) object
  - log(string)
  - sleep(int)

If one of this returns `undefined` instead of the specified type, it means
that it errored out.