Refact stdin to make it re-usable
It is not possible to close os.Stdin and stop reads. So the logic
is that the goroutine is started only once when the module is executed.
Then every time the module stdin is started, it'll read from datac.
If the goroutine is done reading from stdin, it closes datac.
If datac is closed, it closes cancel. Then the main loop now also
selects on cancel which incidentally will close the pipeline.