Add metadata feature (#71)
This metadata feature adds a way to use go templates to templatize the flags.

The way it works is that each new channel sends the callback. Then you
have to call the `Start(map[string]interface{})` method from the `MessageChannel`.
By doing so, it'll pass the metadata to the next module when the module will get it
from the main communication's channel.

Therefor, you should always call `Start()` first then the received callback.