New TLS module
This new module acts as an TLS endpoint and should be used if you need
a TLS server.
It has two modes:
  - bypass: Relay the TLS handshake down to the pipeline. It is expected
      that the rest of the handshake comes back to the module so it can
      be finished. In this mode, the data still stays encrypted, only
      the handshake is visible to the module.
  - intercept: Perform Man in the Middle attack by generating a certificate
      on the fly and presenting this certificate to the client.
      The certificate is signed by a CA that is configured when the module
      starts.

The `decrypt` flag controls this behavior. It is a template that is computed
before the metadata (so it can pass the `decrypt` value to other modules).
It needs to either return `false` or `true`. Any other value is considered
and errors and the channel is closed.

CA: if you don't specify `--ca-cert` and `--ca-key` and if you have `--decrypt`
set, it will generate a certificate key pair for you. You can save that
certificate if you need to reuse it next run.